chore(pushup): move module to data and domain layer
