+++ data/src/main/java/com/example/alpacamusclemaintenance/data/pushup/PushUpRepository.kt
@@ -1,0 +1,12 @@
+package com.example.alpacamusclemaintenance.data.pushup
+
+import androidx.lifecycle.LiveData
+import com.example.alpacamusclemaintenance.domain.pushup.PushUp
+import javax.inject.Inject
+
+interface PushUpRepository {
+
+    fun getPushUps(): LiveData<List<PushUp>>
+
+    suspend fun insert(pushUp: PushUp)
+}
